fix(upload): return rejections on partial additional-image batches
Mixed admin uploads (some valid, some decode-failed) now include
rejections[] on 200 responses, not only when every file fails.

Extract shared helpers into additional_image_upload.py,
turtle_folder_images.py, and review_packet_format.py; add unit and
integration tests for partial-batch behavior.
