+++ backend/routes/review.py
@@ -3,7 +3,6 @@ Review queue endpoints
 """
 
 import os
-import json
 import time
 import traceback
 from typing import Optional, Tuple
@@ -16,13 +15,21 @@ from services.manager_service import get_sheets_service, get_community_sheets_se
 from config import UPLOAD_FOLDER, MAX_FILE_SIZE, allowed_file
 from image_utils import UploadImageError
 from upload_rate_limit import upload_rate_limit_ok, upload_rate_limit_response
-from upload_validation import ingest_saved_upload, log_upload_rejection
+from upload_validation import ingest_saved_upload
 from turtle_manager import canonical_new_turtle_folder_id  # re-exported for callers/tests
 from general_locations_catalog import (
     get_sheet_default,
     resolve_general_location_from_sheet_and_value,
 )
-from additional_image_labels import normalize_additional_type, normalize_label_list, parse_labels_from_form
+from additional_image_upload import (
+    additional_upload_success_json,
+    cleanup_temp_upload_paths,
+    collect_indexed_additional_uploads,
+    no_valid_files_json,
+)
+from review_packet_format import format_review_packet_item
+
+__all__ = ['format_review_packet_item', 'normalize_new_turtle_location_for_disk', 'canonical_new_turtle_folder_id']
 
 
 def normalize_new_turtle_location_for_disk(
@@ -92,138 +99,6 @@ def normalize_new_turtle_location_for_disk(
     return (out, resolved_general_loc)
 
 
-def format_review_packet_item(packet_dir, request_id):
-    """Build one queue item dict from packet_dir (used by get_review_queue and get_review_packet)."""
-    metadata_path = os.path.join(packet_dir, 'metadata.json')
-    metadata = {}
-    if os.path.exists(metadata_path):
-        with open(metadata_path, 'r') as f:
-            metadata = json.load(f)
-
-    additional_images = []
-    additional_dir = os.path.join(packet_dir, 'additional_images')
-
-    def parse_manifest_or_folder(target_dir):
-        results = []
-        manifest_path = os.path.join(target_dir, 'manifest.json')
-        processed_files = set()
-
-        if os.path.isfile(manifest_path):
-            try:
-                with open(manifest_path, 'r') as f:
-                    manifest = json.load(f)
-                for entry in manifest:
-                    fn = entry.get('filename')
-                    kind = entry.get('type', 'other')
-                    if fn:
-                        p = os.path.join(target_dir, fn)
-                        if os.path.isfile(p):
-                            row = {
-                                'filename': fn,
-                                'type': kind,
-                                'timestamp': entry.get('timestamp'),
-                                'image_path': p,
-                            }
-                            lbs = entry.get('labels')
-                            if lbs:
-                                row['labels'] = normalize_label_list(lbs)
-                            results.append(row)
-                            processed_files.add(fn)
-            except (json.JSONDecodeError, OSError):
-                pass
-
-        if os.path.isdir(target_dir):
-            for f in sorted(os.listdir(target_dir)):
-                if f != 'manifest.json' and f not in processed_files and f.lower().endswith(('.jpg', '.jpeg', '.png', '.gif', '.webp')):
-                    results.append({
-                        'filename': f,
-                        'type': 'other',
-                        'labels': [],
-                        'timestamp': None,
-                        'image_path': os.path.join(target_dir, f),
-                    })
-        return results
-
-    if os.path.isdir(additional_dir):
-        additional_images.extend(parse_manifest_or_folder(additional_dir))
-        for item in sorted(os.listdir(additional_dir)):
-            item_path = os.path.join(additional_dir, item)
-            if os.path.isdir(item_path):
-                additional_images.extend(parse_manifest_or_folder(item_path))
-
-    uploaded_image = None
-    for f in os.listdir(packet_dir):
-        if f.lower().endswith(('.jpg', '.png', '.jpeg')) and f != 'metadata.json' and not f.startswith('.'):
-            uploaded_image = os.path.join(packet_dir, f)
-            break
-
-    candidates_dir = os.path.join(packet_dir, 'candidate_matches')
-    failed_path = os.path.join(packet_dir, 'match_search_failed.json')
-    match_search_failed = os.path.isfile(failed_path)
-    match_search_error = None
-    if match_search_failed:
-        try:
-            with open(failed_path, 'r', encoding='utf-8') as f:
-                fail_data = json.load(f)
-            if isinstance(fail_data, dict):
-                err = (fail_data.get('error') or '').strip()
-                match_search_error = err or None
-        except (json.JSONDecodeError, OSError, TypeError):
-            pass
-        if match_search_error is None:
-            match_search_error = 'Match search failed.'
-
-    # Staff/admin /api/upload builds the packet synchronously before search. If search errors,
-    # candidate_matches is never created; without a failure file (older uploads), the folder
-    # is not "still matching" — classify as failed so the queue shows recovery actions.
-    if (
-        request_id.startswith('admin_')
-        and not os.path.isdir(candidates_dir)
-        and not match_search_failed
-    ):
-        match_search_failed = True
-        match_search_error = (
-            'Match search did not complete for this staff/admin upload. '
-            'Try uploading again, or create a new turtle from this find.'
-        )
-
-    # candidate_matches is created after SuperPoint search succeeds in create_review_packet;
-    # missing dir + no failure marker => matching still running (or legacy stuck community packet).
-    match_search_pending = not os.path.isdir(candidates_dir) and not match_search_failed
-    candidates = []
-    if os.path.isdir(candidates_dir):
-        for candidate_file in sorted(os.listdir(candidates_dir)):
-            if candidate_file.lower().endswith(('.jpg', '.png', '.jpeg')):
-                # Strip extension case-insensitively (.JPG was left on parts, breaking int() for Rank/Conf).
-                root, ext = os.path.splitext(candidate_file)
-                base_name = root if ext.lower() in ('.jpg', '.jpeg', '.png') else candidate_file
-                parts = base_name.split('_')
-                rank, turtle_id, confidence = 0, 'Unknown', 0
-                for part in parts:
-                    if part.startswith('Rank'):
-                        rank = int(part.replace('Rank', ''))
-                    elif part.startswith('ID'):
-                        turtle_id = part.replace('ID', '')
-                    elif part.startswith('Conf'):
-                        confidence = int(part.replace('Conf', ''))
-                    elif part.startswith('Score'):
-                        confidence = 0
-                candidates.append({'rank': rank, 'turtle_id': turtle_id, 'confidence': confidence, 'image_path': os.path.join(candidates_dir, candidate_file)})
-
-    return {
-        'request_id': request_id,
-        'uploaded_image': uploaded_image,
-        'metadata': metadata,
-        'additional_images': additional_images,
-        'candidates': sorted(candidates, key=lambda x: x['rank']),
-        'match_search_pending': match_search_pending,
-        'match_search_failed': match_search_failed,
-        'match_search_error': match_search_error,
-        'status': 'pending',
-        'photo_type': metadata.get('photo_type', 'plastron'),
-    }
-
-
 def register_review_routes(app):
     """Register review queue routes"""
     
@@ -259,105 +134,23 @@ def register_review_routes(app):
         if manager_service.manager is None:
             return jsonify({'error': 'TurtleManager failed to initialize'}), 500
         files_with_types = []
-        rejections = []
         try:
-            for key in list(request.files.keys()):
-                if not key.startswith('file_'):
-                    continue
-                f = request.files[key]
-                if not f or not f.filename:
-                    continue
-                idx = key.replace('file_', '')
-                typ = normalize_additional_type(request.form.get(f'type_{idx}'))
-                lbs = parse_labels_from_form(request.form, idx)
-                if not allowed_file(f.filename):
-                    msg = 'Invalid file type. Allowed: JPEG, PNG, GIF, WEBP, HEIC.'
-                    rejections.append({
-                        'filename': f.filename,
-                        'code': 'invalid_extension',
-                        'error': msg,
-                    })
-                    log_upload_rejection(
-                        context='review/additional',
-                        path=None,
-                        code='invalid_extension',
-                        message=msg,
-                        filename=f.filename,
-                    )
-                    continue
-                f.seek(0, os.SEEK_END)
-                size = f.tell()
-                f.seek(0)
-                if size > MAX_FILE_SIZE:
-                    msg = 'File too large (max 8MB after optimization).'
-                    rejections.append({
-                        'filename': f.filename,
-                        'code': 'file_too_large',
-                        'error': msg,
-                    })
-                    log_upload_rejection(
-                        context='review/additional',
-                        path=None,
-                        code='file_too_large',
-                        message=msg,
-                        filename=f.filename,
-                    )
-                    continue
-                orig_safe = secure_filename(f.filename) or ''
-                ext = os.path.splitext(orig_safe)[1] or '.jpg'
-                temp_path = os.path.join(UPLOAD_FOLDER, f"review_extra_{request_id}_{idx}_{int(time.time())}{ext}")
-                f.save(temp_path)
-                try:
-                    temp_path = ingest_saved_upload(
-                        temp_path, context='review/additional', filename=f.filename,
-                    )
-                except UploadImageError as img_err:
-                    rejections.append({
-                        'filename': f.filename,
-                        'code': img_err.code,
-                        'error': img_err.message,
-                    })
-                    if os.path.isfile(temp_path):
-                        try:
-                            os.remove(temp_path)
-                        except OSError:
-                            pass
-                    continue
-                item = {
-                    'path': temp_path,
-                    'type': typ,
-                    'timestamp': time.strftime('%Y-%m-%dT%H:%M:%SZ', time.gmtime()),
-                    'original_filename': os.path.basename(orig_safe) if orig_safe else f'upload{ext}',
-                }
-                if lbs:
-                    item['labels'] = lbs
-                files_with_types.append(item)
+            files_with_types, rejections = collect_indexed_additional_uploads(
+                request,
+                context='review/additional',
+                temp_path_for_index=lambda idx, ext: os.path.join(
+                    UPLOAD_FOLDER, f"review_extra_{request_id}_{idx}_{int(time.time())}{ext}",
+                ),
+            )
             if not files_with_types:
-                primary = rejections[0] if rejections else None
-                return jsonify({
-                    'error': primary['error'] if primary else 'No valid image files provided',
-                    'code': primary['code'] if primary else 'no_valid_files',
-                    'rejections': rejections,
-                }), 400
+                return jsonify(no_valid_files_json(rejections)), 400
             success, msg = manager_service.manager.add_additional_images_to_packet(request_id, files_with_types)
-            for item in files_with_types:
-                p = item.get('path')
-                if p and os.path.isfile(p):
-                    try:
-                        os.remove(p)
-                    except OSError:
-                        pass
+            cleanup_temp_upload_paths(files_with_types)
             if not success:
                 return jsonify({'error': msg}), 400
-            return jsonify({'success': True, 'message': f'Added {len(files_with_types)} image(s).'})
+            return jsonify(additional_upload_success_json(files_with_types, rejections))
         except Exception as e:
-            for item in files_with_types:
-                p = item.get('path')
-                if p and os.path.isfile(p):
-                    try:
-                        os.remove(p)
-                    except OSError:
-                        pass
+            cleanup_temp_upload_paths(files_with_types)
             return jsonify({'error': str(e)}), 500
 
     @app.route('/api/review-queue/<request_id>/additional-images', methods=['DELETE'])
